Various fixes
- Made the project compile again with a more recent version of monogame.
Added some dummy Effect classes for this reason.
- Made the project easier to work with when using it within a Visual
Studio Tools for Unity solution. Changed the configuration names from
Unity-Debug to just Debug because Visual Studio Tools for Unity changes
this to Debug every time the solution file is regenerated, causing the
XnaUnitySpriteBatch project not to be build at all. Added some post
build commands to copy the build dlls to the unity project.
- Fixed some issues in GraphicsDevice. "Custom/SpriteShader" isn't
included in the repo so for now the "Sprites/Default" shader is hooked
in. Now meshes drawn using MeshHolder.Populate aren't culled too soon
anymore by recalculating the bounds on the mesh.