Reworked the audio system.
Now the complete MonoGame audio system is used with corresponding
platform specific partial classes for Unity.
So now most of the feature set of XNA/MonoGame is available, the
UnityEngine.AudioSource objects are pooled using the MonoGame
SoundEffectInstancePool.
MediaPlayer.IsRepeating has been added, which is the only extra part of
the api I need right now.