Fixed memory leak by erroneous creation of new meshes.
An off by one error in a variable tracking the number of buffers in the GraphicsDevice.buffers array caused the last element to not be restarted properly. This caused nextFrontMesh to not be reset to 0, which caused other code to assume the buffer was too small and growing it, every single frame.