Made the scene texture freeing form memory more efficient.
We used to call UnityEngine.Resources.UnloadUnusedAssets() multiple times (once for each bundle)
when exiting a game. This took a long while, probably because it's an async process so having multiple
of those processes running, all locking the same stuff, may not help the situation.
No for each bundle we only request a UnloadUnusedAssets call and actually do it one second after
the last request came through.